+++ frontend/package.json
@@ -18,6 +18,7 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "supabase": "^2.64.2",
     "tailwindcss": "^4",
     "typescript": "^5"
   }
