+++ frontend/src/data/index.ts
@@ -1,0 +1,5 @@
+// Data Layer - Interface Adapters (Repositories/Gateways)
+// This layer implements the repository interfaces defined in the domain layer
+
+export * from "./repositories";
+
