+++ frontend/src/data/repositories/index.ts
@@ -1,0 +1,2 @@
+export { BaseballRepositoryImpl } from "./BaseballRepositoryImpl";
+
