+++ frontend/src/domain/entities/Game.ts
@@ -1,0 +1,14 @@
+import { PlayerSummary } from "./Player";
+
+export type TeamInfo = {
+  code: string;
+  name: string;
+};
+
+export type GameAnalysisResponse = {
+  game_id: number;
+  game_date: string;
+  teams: TeamInfo[];
+  players: PlayerSummary[];
+};
+
