+++ frontend/src/domain/entities/index.ts
@@ -1,0 +1,4 @@
+export { PlayerStats, PlayerSummary } from "./Player";
+export { TeamInfo, GameAnalysisResponse } from "./Game";
+export { HeatmapPoint, SwingPoint, PlayerHeatmapResponse } from "./Heatmap";
+
