+++ frontend/src/domain/index.ts
@@ -1,0 +1,8 @@
+// Domain Layer - Core Business Logic
+// This is the innermost layer containing enterprise business rules
+
+export * from "./entities";
+export * from "./repositories";
+export * from "./usecases";
+export * from "./constants";
+
