+++ frontend/src/domain/repositories/index.ts
@@ -1,0 +1,2 @@
+export { IBaseballRepository } from "./IBaseballRepository";
+
