+++ frontend/src/domain/usecases/index.ts
@@ -1,0 +1,3 @@
+export { GetGameAnalysisUseCase } from "./GetGameAnalysisUseCase";
+export { GetPlayerHeatmapUseCase } from "./GetPlayerHeatmapUseCase";
+
