+++ frontend/src/presentation/components/index.ts
@@ -1,0 +1,2 @@
+export { StrikeZoneHeatmap, type StrikeZoneHeatmapRef } from "./StrikeZoneHeatmap";
+
