+++ frontend/src/presentation/constants.ts
@@ -1,0 +1,2 @@
+export const FALLBACK_HEADSHOT = "https://via.placeholder.com/80x80?text=?";
+
