+++ frontend/src/presentation/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useDashboard } from "./useDashboard";
+
