+++ frontend/src/presentation/index.ts
@@ -1,0 +1,8 @@
+// Presentation Layer - UI (Views), Controllers, Presenters, ViewModels
+// This is the outermost layer containing UI components and presentation logic
+
+export * from "./components";
+export * from "./hooks";
+export * from "./pages";
+export * from "./constants";
+
