+++ frontend/src/presentation/pages/index.ts
@@ -1,0 +1,3 @@
+export { default as HomePage } from "./HomePage";
+export { default as DashboardPage } from "./DashboardPage";
+
