--- package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "supabase": "^2.48.3"
-  }
-}
